docs: refactor accordion demos to use new design
