+++ docs/app/docs/components/[slug]/components/accordion/basic.demo.tsx
@@ -4,10 +4,10 @@ import {
   For,
   Show,
 } from '@cerberus-design/react'
+import Image from 'next/image'
+import { Suspense } from 'react'
 import { Box } from 'styled-system/jsx'
 import data from './data.json'
-
-export const BASIC = 'accordion.basic'
 
 export function BasicDemo() {
   return (
@@ -20,11 +20,28 @@ export function BasicDemo() {
               key={item.id}
               value={item.value}
             >
-              <Show when={item.content !== null}>{item.content}</Show>
+              <Show when={item.content !== null} fallback={<FallbackContent />}>
+                {item.content}
+              </Show>
             </AccordionItemGroup>
           )}
         </For>
       </Accordion>
+    </Box>
+  )
+}
+
+function FallbackContent() {
+  return (
+    <Box position="relative" w="full">
+      <Suspense>
+        <Image
+          alt="A fiery cerberus"
+          src="https://images-wixmp-ed30a86b8c4ca887773594c2.wixmp.com/f/4d7b49a9-2590-4dda-88ad-8046da56428b/dg97zza-65055527-eaf3-48e6-86aa-cf70a0880eea.png?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1cm46YXBwOjdlMGQxODg5ODIyNjQzNzNhNWYwZDQxNWVhMGQyNmUwIiwiaXNzIjoidXJuOmFwcDo3ZTBkMTg4OTgyMjY0MzczYTVmMGQ0MTVlYTBkMjZlMCIsIm9iaiI6W1t7InBhdGgiOiJcL2ZcLzRkN2I0OWE5LTI1OTAtNGRkYS04OGFkLTgwNDZkYTU2NDI4YlwvZGc5N3p6YS02NTA1NTUyNy1lYWYzLTQ4ZTYtODZhYS1jZjcwYTA4ODBlZWEucG5nIn1dXSwiYXVkIjpbInVybjpzZXJ2aWNlOmZpbGUuZG93bmxvYWQiXX0.BT7UC12eP5ibJs6uEYPIOQNQ6Xdnvw1ttIrCEgPvcSk"
+          height={447}
+          width={794}
+        />
+      </Suspense>
     </Box>
   )
 }
