+++ docs/app/docs/components/[slug]/content/accordion.mdx
@@ -9,8 +9,7 @@ ark: 'accordion'
 
 import CodePreview from '@/app/components/CodePreview'
 import { NoteAdmonition } from '@/app/components/Admonition'
-import { BasicDemo, BASIC } from '../components/accordion/basic.demo'
-import { StaticPreview, IndicatorPositionPreview, SizesPreview, CustomPreview } from '../components/accordion/static'
+import { DEMOS } from '../components/accordion/static'
 
 ```ts
 import {
@@ -23,68 +22,19 @@ import {
 
 The Accordion component is a controlled abstraction of the primitive components that can be used to show or hide content. It can be used in a group with other accordion items to allow for multiple selections.
 
-<CodePreview id={BASIC} preview={<BasicDemo />} />
+<CodePreview {...DEMOS.basic} />
 
 ## Indicator Position
 
 The `AccordionItemGroup` component supports two indicator positions: `start` and `end`. This changes the position of the icon indicator displayed.
 
-<CodePreview preview={<IndicatorPositionPreview />}>
-```tsx title="indicator position demo" {11}
-function BasicPreview() {
-  return (
-    <Box w="2/3">
-      <Accordion defaultValue={['one']}>
-        <For each={data}>
-          {(item) => (
-            <AccordionItemGroup
-              heading={item.heading}
-              key={item.id}
-              value={item.value}
-              indicatorPosition="start"
-            >
-              <Show when={item.content !== null} fallback={<FallbackContent />}>
-                {item.content}
-              </Show>
-            </AccordionItemGroup>
-          )}
-        </For>
-      </Accordion>
-    </Box>
-  )
-}
-```
-</CodePreview>
+<CodePreview {...DEMOS.indicator} />
 
 ## Sizes
 
 The `Accordion` component supports two sizes: `sm` and `lg`. This changes the size of the icon indicator displayed.
 
-<CodePreview preview={<SizesPreview />}>
-```tsx title="sizes demo"
-function BasicPreview() {
-  return (
-    <Box w="2/3">
-      <Accordion defaultValue={['one']} size="sm">
-        <For each={data}>
-          {(item) => (
-            <AccordionItemGroup
-              heading={item.heading}
-              key={item.id}
-              value={item.value}
-            >
-              <Show when={item.content !== null} fallback={<FallbackContent />}>
-                {item.content}
-              </Show>
-            </AccordionItemGroup>
-          )}
-        </For>
-      </Accordion>
-    </Box>
-  )
-}
-```
-</CodePreview>
+<CodePreview {...DEMOS.size} />
 
 ## Primitives
 
@@ -98,55 +48,7 @@ You can utilize the primitive components or the `css` prop to customize the acco
 | AccordionItemContent    | The content that is shown or hidden when the trigger is activated. |
 | AccordionItemIndicator  | An optional indicator to show the state of the accordion item. |
 
-<CodePreview preview={<CustomPreview />}>
-```tsx title="custom accordion"
-import { AccordionParts } from '@cerberus/react'
-import { Box } from 'styled-system/jsx'
-import { css } from 'styled-system/css'
-
-export function CustomPreview() {
-  return (
-    <Box w="2/3">
-      <AccordionParts.Root
-        css={{
-          transform: 'skewX(-10deg)',
-        }}
-      >
-        <AccordionParts.Item
-          css={{
-            bgColor: 'black',
-            borderColor: 'red',
-          }}
-          value="one"
-        >
-          <AccordionParts.ItemTrigger
-            css={{
-              color: 'red',
-            }}
-            data-indicator-position="start"
-          >
-            <AccordionParts.ItemIndicator>🔥</AccordionParts.ItemIndicator>
-            Cerberus
-          </AccordionParts.ItemTrigger>
-
-          <AccordionParts.ItemContent
-            css={{
-              color: 'white',
-              paddingInline: 'md',
-              textStyle: 'body-md',
-            }}
-          >
-            Cerberus is the three-headed dog that guards the gates of the
-            Underworld and our sweet baby boi protecting the integrity of your
-            design system.
-          </AccordionParts.ItemContent>
-        </AccordionParts.Item>
-      </AccordionParts.Root>
-    </Box>
-  )
-}
-```
-</CodePreview>
+<CodePreview {...DEMOS.custom} />
 
 ## API
